requirements.txt
asgiref==3.2.10
beautifulsoup4==4.8.2
braintree==4.2.0
certifi==2020.12.5
chardet==3.0.4
defusedxml==0.6.0
Django==3.1
django-address==0.2.5
django-admin-parent-filter==2020.12.3
django-allauth==0.42.0
django-debug-toolbar==2.2
django-el-pagination==3.3.0
django-extensions==3.0.5
django-filter==2.3.0
django-modelcluster==5.0.2
django-recaptcha==2.0.6
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2020.6.1
oauthlib==3.1.0
openpyxl==3.0.5
Pillow==7.2.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
Unidecode==1.1.2
urllib3==1.25.10
wagtail==2.10.1
wagtail-cache==1.0.1
wagtail-django-recaptcha==1.0
wagtail-orderable==1.0.3
webencodings==0.5.1
Werkzeug==1.0.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
google-cloud-secret-manager==2.0.0
google-auth==1.22.1
django-storages[google]==1.9.1
django-environ==0.4.5
